% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discriminantMDS.R
\name{as.data.frame.LinearClassifier}
\alias{as.data.frame.LinearClassifier}
\title{Combine predictions from a LinearClassifier with cData from the object used to generate it}
\usage{
\method{as.data.frame}{LinearClassifier}(lc, row.names, optional, ...)
}
\arguments{
\item{lc}{LinearClassifier}

\item{row.names}{ignored}

\item{optional}{ignored}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Combine predictions from a LinearClassifier with cData from the object used to generate it
}

