% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_check_paths.R
\name{CheckPaths}
\alias{CheckPaths}
\title{Validate barcode inputs}
\usage{
CheckPaths(input)
}
\description{
CheckPaths will verify that the specified input path contains the necessary
barcode matrices, and then return a console output to verify that all files
were found. Will produce an error if the path is invalid or a file is missing.
}
