% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{cv.glmnet,formula-method}
\alias{cv.glmnet,formula-method}
\title{Generic for cv.glmnet}
\usage{
\S4method{cv.glmnet}{formula}(x, y, ...)
}
\arguments{
\item{x}{formula}

\item{y}{data.frame or environment in which \code{x} is evaluated}

\item{...}{arguments passed to cv.glmnet}
}
\value{
see cv.glmnet
}
\description{
Accepts formula arguments
}

