% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{doGLMnet}
\alias{doGLMnet}
\title{Evaluate a glmSingleCellAssay classifier}
\usage{
doGLMnet(train, predict, s = "lambda.1se")
}
\arguments{
\item{train}{output from \code{glmSingleCellAssay}, used for its fitted coefficients}

\item{predict}{output from \code{glmSingleCellAssay}, used for its model.matrix to predict (possibly) out of sample}

\item{s}{glmnet penalty parameter, default 'lambda.1se'}
}
\value{
object of class 'Linear Classifier' and 'matrix'
}
\description{
Wraps a glmnet classifier (derived from glmSingleCellAssay) to get the scores (coefficients) and predictions.
}
\seealso{
\link{doLDA}, \link{annotateBiPlot}, \link{glmSingleCellAssay}
}

