% Generated by roxygen2 (4.0.2): do not edit by hand
\name{doLDA}
\alias{doLDA}
\title{Run Linear Discriminant Analysis}
\usage{
doLDA(train, predict, grouping, Layer = "lCount")
}
\arguments{
\item{train}{\code{SingleCellAssay}}

\item{predict}{\code{SingleCellAssay}}

\item{grouping}{\code{character} vector of length 1 naming a column in the \code{cData} of 'train' and 'predict'}

\item{Layer}{which layer should be used to train and predict?}
}
\value{
object of class 'LinearClassifier' and 'matrix'
}
\description{
Runs Fisher's LDA to discriminate 'grouping' via 'train',
optionally attempting out-of-sample prediction using sample 'predict'
}
\seealso{
\link{doGLMnet}, \link{annotateBiPlot}
}

