% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitZifNetwork}
\alias{fitZifNetwork}
\title{Fit Meinhousen-Buhlmann to a SingleCellAssay object}
\usage{
fitZifNetwork(sc, additive.effects, min.freq = 0.05,
  gene.predictors = "zero.inflated", precenter = TRUE,
  precenter.fun = scale, response = "hurdle", modelSelector,
  onlyReturnFitter = FALSE, debug = FALSE, ...)
}
\arguments{
\item{sc}{SingleCellAssay object on a thresholded layer}

\item{additive.effects}{character vector, possibly using formula syntax of columns from \code{cData(sc)} to be included as unpenalized terms.}

\item{min.freq}{genes below this frequency are excluded as predictors and dependent variables}

\item{gene.predictors}{ignored}

\item{precenter}{How should centering/scaling be done with respect to continuous regressions.  TRUE if centering should be done with respect to all cells; FALSE if centering should be done only with respect to expressed cells
When precenter=TRUE, cv.glmnet will not standardize.}

\item{precenter.fun}{a function called to center the expression matrix prior to calling glmnet}

\item{response}{a character vector, one of 'zero.inflated', 'hurdle', or 'cg.regression', 'cg.regression2', 'cg.mle'}

\item{modelSelector}{a function called gene and component-wise on each fit, that should return an index to the glmnet lambda sequence for that gene and component}

\item{onlyReturnFitter}{if TRUE, return an undocumented fitter function that is internally called on each gene/component.}

\item{...}{passed to cv.glmnet}
}
\value{
2-D list of cv.glmnet objects with attributes
}
\description{
Regresses the dichotomous and continuous components of each gene in \code{sc}
on every other gene.
\code{additive.effects} from \code{cData(sc)} are included unpenalized.
Currently only \code{gene.predictors} as 'zero.inflated' is supported.
... is passed along to cv.glmnet, see documentation there.
}

