% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fortify.zifnetwork}
\alias{fortify.zifnetwork}
\title{Derive global properties of network fit}
\usage{
fortify.zifnetwork(fits, lc.range, ld.range, nknots = 20, ebic.lambda = 0)
}
\arguments{
\item{fits}{\code{fitZifNetwork} object}

\item{lc.range}{optional}

\item{ld.range}{optional}

\item{nknots}{}

\item{ebic.lambda}{lambda penalty for extended Bayesian Info Crit. (Rigel and Drton)}
}
\value{
list with entries \code{fortified}, \code{norm.grid}, \code{native.path}
fortified and native.path are both data.frames with entries for each primerid, containing statistics of the fit as lambda varies.
fortified has cartesian product of continuous lambda values and discrete, over the same grid for each gene.
native.path has
}
\description{
.. content for \details{} ..
}

