% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{getZifNetwork}
\alias{getZifNetwork}
\title{Put coefficients from a set of network regressions into a matrix/array}
\usage{
getZifNetwork(listOfFits, l.c, l.d, collapse = FALSE, union = TRUE, layers)
}
\arguments{
\item{listOfFits}{output from fitZifNetwork}

\item{l.c}{continuous lambda value}

\item{l.d}{discrete lambda value}

\item{collapse}{should the network be collapsed between layers?}

\item{constraint}{a constraint, which will be translated into the regularization tuning parameter lambda for each regression}
}
\value{
an array
}
\description{
The array is ngenes X ngenes X {2,4}, with the last dimension depending
on whether zero-inflated or hurdle predictors were used.
When type = 'grp.l1' lambda for each regression is select to minimize the hurdle deviance subject to the grouped l1 norm being less than 'constraint'.
When type = 'comb.l1', ditto but with the regular l1 norm being constrained.
When type = 'lambda.min' lambda is selected to be 'constraint' times the lambda.min determined by cross validation.
}

