% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glmSingleCellAssay}
\alias{glmSingleCellAssay}
\title{##Generic for cv.glmnet
##
##Accepts formula arguments
##}
\usage{
glmSingleCellAssay(sca, comparison, min.freq, predictor = c("continuous",
  "dichotomous"), pen.scale.interaction = 2, precenter = FALSE,
  prescale = FALSE, addn = NULL, addn.penalty = 0, user.mm, alpha = 0.9,
  only.mm = FALSE, ...)
}
\arguments{
\item{x}{formula
##}

\item{y}{data.frame or environment in which \code{x} is evaluated
##}

\item{...}{arguments passed to cv.glmnet
##}

\item{sca}{SingleCellAssay object}

\item{comparison}{character naming a column in \code{cData(sca)}}

\item{min.freq}{minimum frequency for a gene to be considered in the classifier}

\item{predictor}{character vector naming some combination of 'continuous', 'dichotomous' or 'interaction'.  See details.}

\item{pen.scale.interaction}{multiply the l1 penalty by this factor if interactions are included}

\item{precenter}{should the gene predictors be centered? Recommended if there are interactions present to reduce co-linearity of the interaction with the marginal term.}

\item{prescale}{should the gene predictors be scaled to have unit variance?}

\item{addn}{character vector, giving additional columns of design, interpreted in the context of cData(sca)}

\item{addn.penalty}{an optional numeric giving the relative scale of the penalty for add}

\item{user.mm}{a function to be applied to exprs(sca) instead of the defaults given by \code{predictor}}

\item{alpha}{elasticnet penalty parameter. Default =.9.}

\item{only.mm}{Should only the model matrix be returned, rather than actually calling cv.glmnet?}

\item{...}{additional arguments to cv.glmnet.}
}
\value{
see cv.glmnet
Run a multinomial lasso on a SingleCellAssay object to predict group membership

This function generates a design matrix based on the expression values in \code{sca}
and calls \code{cv.glmnet} to try to classify a group named by \code{comparison}, which keys a column in the \code{cData} of \code{sca}

The design matrix is generated according to the option \code{predictor}. If \code{predictor} vector includes the term 'dichotomous', then each gene is treated as binary indicators.  If the term 'continuous' is included, then the zero-inflated (continuous) value for the gene is used.  If both 'continuous' and 'dichotomous' are included, then the both values for the gene are used, however the continuous values are centered about their conditional mean using the function \code{xform}.  If 'interaction' is included, then all the terms are crossed with each other to generate pairwise interactions.

list with components 'cv.fit' giving the output from cv.glmnet, 'mm' giving the model matrix, 'response' giving the response vector and 'sca' containing the 'sca' passed as input to the function
}
\description{
##Generic for cv.glmnet
##
##Accepts formula arguments
##
}
\seealso{
glmMisclass, getNZdesign, doGLMnet, cv.glmnet
}

