% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{layoutZifNetwork}
\alias{layoutZifNetwork}
\title{Turn an object from fitZifNetwork into a igraph object}
\usage{
layoutZifNetwork(zifFit, Vattr = NULL, Eattr = NULL, collapse = TRUE,
  weight = FALSE, union = TRUE, ...)
}
\arguments{
\item{zifFit}{fitZifNetwork output}

\item{Vattr}{character vector named with vertex names.}

\item{Eattr}{data.frame with columns 'X1' 'X2' naming vertex pairs and 'value' giving the value for this set of edges}

\item{collapse}{should the discrete/continuous parts be collapsed?}

\item{union}{within each layer, should we take the union of a vertex's neighborhood, or should we take the intersection?}

\item{weight}{currently ignored}

\item{...}{passed to getZifNetwork}
}
\value{
igraph object, with attributes
}
\description{
Edges/Vertex may be given colors by setting Vattr or Ettr
}

