% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(raw_dir, filtered_dir)
}
\arguments{
\item{raw_dir}{Path to the raw expression matrix}

\item{filtered_dir}{Path to the filtered expression matrix}
}
\value{
De-multiplexed protein and RNA expression matrices,
}
\description{
This function takes in directory pointers for the raw and filtered expression
matrices and splits them into the RNA and surface protein components. It
also separates empty droplets from cell-containing droplets and stores them
as separate objects for background normalization downstream.
}
