% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discriminantMDS.R
\name{prepClassifier}
\alias{prepClassifier}
\title{Generate biplot vectors from a 'LinearClassifier'}
\usage{
prepClassifier(lc)
}
\arguments{
\item{lc}{'LinearClassifier'}
}
\value{
'LinearClassifer' with attributes
}
\description{
If there are G discrminant functions and P genes,
Generate  'biNorm', an array of G x G x P, giving the 2-dimensional 2-norm of each gene along each 2-combination of groups
'biGenes', also G x G x P giving the ordering of the genes along the 2-norms,
'norm', the G-dimensional 2-norm of each gene.
}

