% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_remove_background.R
\name{RemoveBackground}
\alias{RemoveBackground}
\title{Remove background noise}
\usage{
RemoveBackground(meta.data, background, adt.madrix, rna.matrix, mt.pct = 0.14)
}
\description{
Remove background noise
}
