% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/panelfun.R
\name{singleCellHeat}
\alias{singleCellHeat}
\title{Plot a heatmap of a SingleCellAssay expression matrix}
\usage{
singleCellHeat(ee.sc, rowFactor, cluster = TRUE, order.by.factor = !cluster,
  colFactor, ...)
}
\arguments{
\item{ee.sc}{expression matrix}

\item{rowFactor}{named list of character vectors to be used as a labelling color schema for the rows.}

\item{cluster}{boolean. Apply hclust to rows and columns.}

\item{order.by.factor}{Sort expression matrix by rowFactors and colFactors.  If cluster is TRUE than cluster within rowFactor/colFactor.}

\item{colFactor}{named list of character vectors to be used as a labelling color schema for the columns.}

\item{...}{passed to heatmap.2}
}
\value{
prints plot to current graphic device
}
\description{
Plot a heatmap of a SingleCellAssay expression matrix (eg, from calling exprs).
Optionally cluster and label rows by rowFactor, columns by colFactor, and/or sort the expression matrix by rowFactor/colFactor.
}

