% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{test.pearson.pairwise}
\alias{test.pearson.pairwise}
\title{Generate a matrix of z-statistics comparing two PairwisePearson objects}
\usage{
test.pearson.pairwise(pp1, pp2)
}
\arguments{
\item{pp1}{First PairwisePearson object}

\item{pp2}{Second PairwisePearson object}
}
\value{
matrix of z-statistics
}
\description{
Generate a matrix of z-statistics comparing two PairwisePearson objects
}

