% Generated by roxygen2 (4.0.1): do not edit by hand
\name{annotateBiPlot}
\alias{annotateBiPlot}
\title{Add bi-plot vectors onto a ggpairs scatter plot of classifiers}
\usage{
annotateBiPlot(ggpairsObj, lc, genesToShow = 5, expand = 1,
  where = "lower", ...)
}
\arguments{
\item{ggpairsObj}{output from a call to 'ggpairs'}

\item{lc}{object of class 'LinearClassifier'}

\item{genesToShow}{number of bi-plot vectors to show?}

\item{expand}{scaling factor of bi-plot vectors}

\item{where}{character vector, one or more of "upper" or "lower"}

\item{...}{additional arguments passed to ggplot}
}
\value{
modified ggpairs object, which can be plotted by evaluating it.
}
\description{
Draws vectors representing the "angle" and "length" the top 'genesToShow' predictors have when trying to discminant objects.
Vectors are drawn on the lower triangle of a ggpairs object (probably plotted by calling ggpairs(as.data.frame(lc), ...)) by matching names in the ggpairs object to the names of the classifier functions in lc.
So this function probably won't work unless ggpairs was called on 'lc'.
There is an arbitrary scaling between the scatter plot and the length of these vectors.  By default the longest vector is scaled to have length of the narrowest plot limit, but this can be adjusted via 'expand'.
}
\seealso{
\link{doLDA}, \link{doGLMnet}
}

