% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{as.data.frame.LinearClassifier}
\alias{as.data.frame.LinearClassifier}
\title{Combine predictions from a LinearClassifier with cData from the object used to generate it}
\usage{
\method{as.data.frame}{LinearClassifier}(lc, optional)
}
\arguments{
\item{lc}{LinearClassifier}

\item{optional}{ignored}
}
\value{
data.frame
}
\description{
Combine predictions from a LinearClassifier with cData from the object used to generate it
}

