\name{fitZifNetwork}
\alias{fitZifNetwork}
\title{Fit Meinhousen-Buhlmann}
\usage{
  fitZifNetwork(sc, additive.effects, min.freq = 0.05,
    gene.predictors = "zero.inflated", alpha = 1,
    precenter = TRUE)
}
\arguments{
  \item{sc}{}

  \item{additive.effects}{}

  \item{min.freq}{}

  \item{gene.predictors}{}

  \item{alpha}{}
}
\value{
  2-D list of cv.glmnet objects with attributes
}
\description{
  .. content for \details{} ..
}

