% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getZifNetwork}
\alias{getZifNetwork}
\title{Put coefficients from a set of network regressions into a matrix/array}
\usage{
getZifNetwork(listOfFits, l.c, l.d, collapse = FALSE, union = TRUE, layers)
}
\arguments{
\item{listOfFits}{output from fitZifNetwork}

\item{l.c}{continuous lambda value}

\item{l.d}{discrete lambda value}

\item{collapse}{should the network be collapsed between layers?}

\item{union}{currently ignored}

\item{layers}{upon which layers in the listOfFits should we operate (eg, discrete, continuous or both)}
}
\value{
an array
}
\description{
The array is ngenes X ngenes X {2,4}, with the last dimension depending
on whether zero-inflated or hurdle predictors were used.
}

