% Generated by roxygen2 (4.0.1): do not edit by hand
\name{oddsratio.pairwise}
\alias{oddsratio.pairwise}
\title{Calculate pairwise gene odds ratios}
\usage{
oddsratio.pairwise(sc, Formula, surrenderFreq = 0.05, useFirth = FALSE,
  lm.hook, MAX_IT = 30)
}
\arguments{
\item{sc}{SingleCellAssay object}

\item{Formula}{a formula used to build the design matrix.  Probably should include the term 'logOR.primer2' (if the pairwise odds ratios are indeed desired).}

\item{surrenderFreq}{any pair of genes in which either is expressed above or below this threshold will be skipped}

\item{useFirth}{should biased-reduced regression be used via brglm.  Slow.}

\item{lm.hook}{a function to be called on the fitted regression}

\item{MAX_IT}{if useFirth==TRUE, how many iterations should be permitted?}
}
\value{
PairwiseOddsRatio class, with slots \code{fits} (2-D list of fit) and \code{genes} (character vector of gene names).
}
\description{
Calculates odds ratios between each pair of genes on the
dichotomous level.
}

