% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discriminantMDS.R
\name{stat_ellipse}
\alias{stat_ellipse}
\title{Draw a confidence ellipse}
\usage{
stat_ellipse(mapping = NULL, data = NULL, geom = "path",
  position = "identity", ...)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{geom}{}

\item{position}{}

\item{...}{}
}
\value{
ggplot2 layer
}
\description{
Draw a confidence ellipse
}

